Revert "Agregando y ordenando"
