+++ server/models/Customers.js
@@ -22,8 +22,6 @@ const customerSchema=new Schema({
         type:String,
         trim:true,
     },
-},{
-    timestamps:true
 });
 
 module.exports=mongoose.model('Customers',customerSchema);
