--- server/routes/user.routes.js
@@ -1,9 +1,0 @@
-//config rutas
-const express=require('express');
-const router=express.Router();
-
-router.get('/',function(req,res){
-    res.send('server conection')
-});
-
-module.exports=router;
