+++ src/index.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './App';
+import App from './App.js';
+import Curso from './Curso.js';
+import CrearCurso from './CrearCurso.js';
+import MiCurso from './MiCurso.js';
 
 ReactDOM.render(<React.StrictMode>
-    <App/>
+    <MiCurso/>
   </React.StrictMode>, document.getElementById("root"));
